requirements.txt
tensorflow_hub
matplotlib
pillow
psutil
quart
uvloop
quart-cors
